+++ public/index.html
@@ -1857,6 +1857,6 @@ Here is a card with filled data</textarea
       </section>
     </main>
 
-    <!-- <script src="../src/main.js"></script> -->
+    <script src="./bundle.js"></script>
   </body>
 </html>
